rename ec form folder
